summ_ai_backend/requirements.txt
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
deepl==1.18.0
Django==5.0.8
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
joblib==1.4.2
nltk==3.8.1
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.8.0
regex==2024.7.24
requests==2.32.3
retrying==1.3.4
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.1
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
uvicorn==0.20.0